npe in stdrace
git-svn-id: svn://192.168.1.10/public/CoffeeMud@18816 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Races/StdRace.java
@@ -962,7 +962,8 @@ public class StdRace implements Race
 			final MOB following=mob.amFollowing();
 			final MOB ultFollow=mob.amUltimatelyFollowing();
 			if((following!=null)
-			&&((!following.isMonster())||(!ultFollow.isMonster())))
+			&&((!following.isMonster())
+				||((ultFollow != null) && (!ultFollow.isMonster()))))
 			{
 				final MOB M=(MOB)mob.copyOf();
 				CMLib.threads().deleteAllTicks(M);
